Move files to new package user
