+++ app/src/androidTest/java/com/github/gogetters/letsgo/activities/mocking/MockUserBundleProvider.kt
@@ -1,7 +1,7 @@
 package com.github.gogetters.letsgo.activities.mocking
 
-import com.github.gogetters.letsgo.database.UserBundle
-import com.github.gogetters.letsgo.database.UserBundleProvider
+import com.github.gogetters.letsgo.database.user.UserBundle
+import com.github.gogetters.letsgo.database.user.UserBundleProvider
 
 class MockUserBundleProvider: UserBundleProvider() {
     override fun getUserBundle(): UserBundle? {
