+++ app/src/main/java/com/github/gogetters/letsgo/activities/MainActivity.kt
@@ -7,7 +7,7 @@ import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import com.github.gogetters.letsgo.R
-import com.github.gogetters.letsgo.database.FirebaseUserBundleProvider
+import com.github.gogetters.letsgo.database.user.FirebaseUserBundleProvider
 import com.google.android.material.bottomnavigation.LabelVisibilityMode
 import kotlinx.android.synthetic.main.activity_main.*
