+++ app/src/main/java/com/github/gogetters/letsgo/database/user/LetsGoUser.kt
@@ -1,7 +1,8 @@
-package com.github.gogetters.letsgo.database
+package com.github.gogetters.letsgo.database.user
 
 import android.graphics.Bitmap
 import android.util.Log
+import com.github.gogetters.letsgo.database.Database
 import com.google.android.gms.tasks.Task
 import com.google.android.gms.tasks.Tasks
 import java.util.*
@@ -34,7 +35,7 @@ class LetsGoUser(val uid: String, val db: Database.Companion = Database) {
      * Verifies if the user exists in our database
      */
     fun requireUserExists(): Task<Unit> {
-        return db.readData(userPath).continueWith {
+        return Database.readData(userPath).continueWith {
             val userExists = it.result.value != null
 
             if (userExists) {
@@ -59,7 +60,7 @@ class LetsGoUser(val uid: String, val db: Database.Companion = Database) {
         )
 
         // Add a new document with user's uid
-        return db.updateData(userPath, userData)
+        return Database.updateData(userPath, userData)
             .addOnSuccessListener {
                 Log.d(tag, "LetsGoUser document added for uid: $uid")
             }
@@ -72,7 +73,7 @@ class LetsGoUser(val uid: String, val db: Database.Companion = Database) {
      * Downloads this User's data to the DB. Returns a task to track progress and etc.
      */
     fun downloadUserData(): Task<Unit> {
-        return db.readData(userPath)
+        return Database.readData(userPath)
             .continueWith {
                 for (attribute in it.result.children) {
                     when (attribute.key) {
@@ -96,7 +97,7 @@ class LetsGoUser(val uid: String, val db: Database.Companion = Database) {
      * Deletes this User's data from the DB. Returns a task to track progress and etc.
      */
     fun deleteUserData(): Task<Void> {
-        return db.deleteData(userPath)
+        return Database.deleteData(userPath)
             .addOnSuccessListener {
                 Log.d(tag, "LetsGoUser successfully deleted!")
             }
@@ -150,8 +151,8 @@ class LetsGoUser(val uid: String, val db: Database.Companion = Database) {
      * Use to remove a friend or delete a friend request.
      */
     fun deleteFriend(otherUser: LetsGoUser): Task<Void> {
-        return db.deleteData("$userFriendsPath/${otherUser.uid}").continueWithTask {
-            db.deleteData("${otherUser.userFriendsPath}/${this.uid}")
+        return Database.deleteData("$userFriendsPath/${otherUser.uid}").continueWithTask {
+            Database.deleteData("${otherUser.userFriendsPath}/${this.uid}")
         }.addOnSuccessListener {
             Log.d(tag, "'Friend' successfully deleted")
         }.addOnFailureListener {
@@ -163,9 +164,9 @@ class LetsGoUser(val uid: String, val db: Database.Companion = Database) {
      * Updates friend status for both users!
      */
     private fun updateFriendStatus(
-        otherUser: LetsGoUser,
-        status1: FriendStatus,
-        status2: FriendStatus
+            otherUser: LetsGoUser,
+            status1: FriendStatus,
+            status2: FriendStatus
     ): Task<Void> {
         return requireUserExists().continueWithTask {
             otherUser.requireUserExists().continueWithTask {
@@ -187,7 +188,7 @@ class LetsGoUser(val uid: String, val db: Database.Companion = Database) {
         val path = "$userFriendsPath/${otherUser.uid}"
         Log.d(tag, "Adding friend data. path: $path\tstatus: $status")
 
-        return db.writeData(path, status.name);
+        return Database.writeData(path, status.name);
     }
 
     //-------------------------------------------------------------------------------------------
@@ -209,7 +210,7 @@ class LetsGoUser(val uid: String, val db: Database.Companion = Database) {
      */
     // Sometimes I use the word connections and friends interchangeably
     fun downloadFriends(): Task<Void> {
-        return db.readData(userFriendsPath).continueWithTask {
+        return Database.readData(userFriendsPath).continueWithTask {
             val friendUids: EnumMap<FriendStatus, ArrayList<String>> =
                 EnumMap(FriendStatus::class.java)
             friends = EnumMap(FriendStatus::class.java)
